requirements/main.txt
aiomysql
prometheus_client
everett[yaml]
faf-replay-parser
